+++ src/ai/backend/manager/api/resource.py
@@ -34,16 +34,16 @@ from ai.backend.manager.services.agent.actions.watcher_agent_stop import Watcher
 from ai.backend.manager.services.container_registry.actions.get_container_registries import (
     GetContainerRegistriesAction,
 )
-from ai.backend.manager.services.groups.actions.usage_per_month import UsagePerMonthAction
-from ai.backend.manager.services.groups.actions.usage_per_period import UsagePerPeriodAction
+from ai.backend.manager.services.group.actions.usage_per_month import UsagePerMonthAction
+from ai.backend.manager.services.group.actions.usage_per_period import UsagePerPeriodAction
 from ai.backend.manager.services.resource_preset.actions.check_presets import (
     CheckResourcePresetsAction,
 )
 from ai.backend.manager.services.resource_preset.actions.list_presets import (
     ListResourcePresetsAction,
 )
-from ai.backend.manager.services.users.actions.admin_month_stats import AdminMonthStatsAction
-from ai.backend.manager.services.users.actions.user_month_stats import UserMonthStatsAction
+from ai.backend.manager.services.user.actions.admin_month_stats import AdminMonthStatsAction
+from ai.backend.manager.services.user.actions.user_month_stats import UserMonthStatsAction
 
 from .auth import auth_required, superadmin_required
 from .exceptions import InvalidAPIParameters
