+++ src/ai/backend/manager/models/gql_models/group.py
@@ -21,15 +21,15 @@ from sqlalchemy.engine.row import Row
 from ai.backend.common.types import ResourceSlot
 from ai.backend.manager.models.rbac import ProjectScope
 from ai.backend.manager.models.user import UserRole
-from ai.backend.manager.services.groups.actions.create_group import CreateGroupAction
-from ai.backend.manager.services.groups.actions.delete_group import (
+from ai.backend.manager.services.group.actions.create_group import CreateGroupAction
+from ai.backend.manager.services.group.actions.delete_group import (
     DeleteGroupAction,
 )
-from ai.backend.manager.services.groups.actions.modify_group import ModifyGroupAction
-from ai.backend.manager.services.groups.actions.purge_group import (
+from ai.backend.manager.services.group.actions.modify_group import ModifyGroupAction
+from ai.backend.manager.services.group.actions.purge_group import (
     PurgeGroupAction,
 )
-from ai.backend.manager.services.groups.types import GroupCreator, GroupData, GroupModifier
+from ai.backend.manager.services.group.types import GroupCreator, GroupData, GroupModifier
 from ai.backend.manager.types import OptionalState, TriState
 
 from ..base import (
