+++ src/ai/backend/manager/services/group/actions/BUILD
