+++ src/ai/backend/manager/services/group/actions/delete_group.py
@@ -1,13 +1,15 @@
+import uuid
 from dataclasses import dataclass
 from typing import Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.users.actions.base import UserAction
+from ai.backend.manager.services.group.actions.base import GroupAction
+from ai.backend.manager.services.group.types import GroupData
 
 
 @dataclass
-class DeleteUserAction(UserAction):
-    email: str
+class DeleteGroupAction(GroupAction):
+    group_id: uuid.UUID
 
     @override
     def entity_id(self) -> Optional[str]:
@@ -19,9 +21,10 @@ class DeleteUserAction(UserAction):
 
 
 @dataclass
-class DeleteUserActionResult(BaseActionResult):
+class DeleteGroupActionResult(BaseActionResult):
+    data: Optional[GroupData]
     success: bool
 
     @override
     def entity_id(self) -> Optional[str]:
-        return None
+        return str(self.data.id) if self.data is not None else None
