+++ src/ai/backend/manager/services/group/actions/modify_group.py
@@ -3,8 +3,8 @@ from typing import Optional, override
 from uuid import UUID
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.groups.actions.base import GroupAction
-from ai.backend.manager.services.groups.types import GroupData, GroupModifier
+from ai.backend.manager.services.group.actions.base import GroupAction
+from ai.backend.manager.services.group.types import GroupData, GroupModifier
 
 
 @dataclass
