+++ src/ai/backend/manager/services/group/actions/usage_per_period.py
@@ -3,14 +3,15 @@ from typing import Any, Optional, override
 from uuid import UUID
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.groups.actions.base import GroupAction
+from ai.backend.manager.services.group.actions.base import GroupAction
 
 
 # TODO: Change to batch action
 @dataclass
-class UsagePerMonthAction(GroupAction):
-    month: str
-    group_ids: Optional[list[UUID]] = None
+class UsagePerPeriodAction(GroupAction):
+    start_date: str
+    end_date: str
+    project_id: Optional[UUID] = None
 
     @override
     def entity_id(self) -> Optional[str]:
@@ -18,11 +19,11 @@ class UsagePerMonthAction(GroupAction):
 
     @override
     def operation_type(self):
-        return "usage_per_month"
+        return "usage_per_period"
 
 
 @dataclass
-class UsagePerMonthActionResult(BaseActionResult):
+class UsagePerPeriodActionResult(BaseActionResult):
     # TODO: Define return type
     result: list[Any]
