+++ src/ai/backend/manager/services/group/types.py
